requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
pydantic==2.9.2
pydantic-settings==2.5.2
sqlalchemy==2.0.35
aiosqlite==0.20.0
asyncpg==0.29.0
alembic==1.13.3
redis==5.1.1
httpx==0.27.2
python-dotenv==1.0.1
python-multipart==0.0.12
Pillow==10.4.0
tenacity==9.0.0
structlog==24.4.0
PyJWT==2.9.0
passlib[bcrypt]==1.7.4
slowapi==0.1.9

